scripts/requirements.txt
psycopg2-binary==2.9.9
pytest==7.4.4
pytest-cov==4.1.0
black==23.12.1
ruff==0.1.9
flake8==7.0.0

